+++ src/query.jl
@@ -6,21 +6,27 @@ Search job by `job.id` in the queue.
 Return `job::Job` if found, `nothing` if not found.
 """
 function job_query_by_id(id::Int64)
+    res = nothing
     wait_for_lock()
+    try
         for job in JOB_QUEUE
             if job.id == id
-                release_lock()
-                return job
+                res = job
+                break
             end
         end
-        for job in JOB_QUEUE_OK
-            if job.id == id
-                release_lock()
-                return job
+        if isnothing(res)
+            for job in JOB_QUEUE_OK
+                if job.id == id
+                    res = job
+                    break
+                end
             end
         end
-    release_lock()
-    return nothing # not found
+    finally
+        release_lock()
+    end
+    return res
 end # function
 
 job_query = job_query_by_id